requirements.txt

pillow
numpy>=2.1
tifffile
matplotlib